Fix: Corrected condition for ProtoLog message decoding error.
The error message for failing to decode a ProtoLog message due to an unavailable message hash was being triggered when the hash *was* available. The condition has been inverted to correctly log the error when the message hash is *not* found in the viewer config file.

Bug: 442776354
Flag: EXEMPT minor bug fix affecting error messages only
Change-Id: I90e8ac47ee36748de13395e8212420f8e4a56ea5