Add config to allow landscape-to-portrait camera treatment.
This config allows dynamic camera compat treatment for landscape cameras, as a part of simulate-reqested-orientation treatment, which rotates and crops camera feed to portrait, and changes camera sensor rotation to portrait.

There is an existing property "camera.enable_landscape_to_portrait" that allows static camera compat treatment for landscape cameras, which is triggered and applied in the Camera framework (as opposed to the dynamic treatment which is triggered in Window Manager). Reusing this property however makes testing with- and without- treatment more difficult, and creates conflicts with the devices that have the old treatment. After testing, one of these configs might be deprecated in favor of another, but this separation is easier to isolate during development.

Flag: com.android.window.flags.camera_compat_landscape_camera_support
Bug: 365725400
Test: Tested on a desktop device, with landscape camera.
Test: atest WmTests:AppCompatCameraSimReqOrientationPolicyTests
Test: atest WmTests:AppCompatCameraOverridesTest
Change-Id: I60491754c1e11f2d1d8520b0411a37ab7c7b7f12