Avoid NPE in rare cases in CTS tests
In some CTS tests, we use a stub MediaRoute2ProviderService and have
it very rapidly cycle between publishing routes, unpublishing routes,
creating sessions, releasing sessions, and tearing down the
provider. Every once and a while in these test scenarios we can end up
processing a create session request when the provider has begun the
process of shutting down, although this is unlikely to happen in real
world usage. This CL fixes an NPE that happens in this case.

Bug: 443591521
Test: atest --rerun-until-failure 100 MediaRoute2ProviderServiceTest SystemMediaRouter2Test
Flag: EXEMPT BUGFIX
Change-Id: I9058ad62468db0e0b4cfc20a0b92f297b5ed70cb