[WindowDecor] Notify WindowDecor only after split state is updated
There is an incorrect order of events when a split screen pair is
updated currently. StageTaskListener notifies WindowDecor of
onTaskInfoChanged, and DesktopModeWindowDecorViewModel receives the
signal and calls relayout. Then, StageTaskListener updates the split
state of current tasks.

This led to a bug where the input layer of the app handle for the
updated task in split is shown in the wrong location, due to
SplitScreenController not having the correct state yet.

This CL moves the onTaskInfoChanged call in StageTaskListener to after
the onChildTaskStatusChanged call is made, which updates the split
state. This ensures that by the time AppHandleController#relayout is
called, SplitScreenController has the correct split state of tasks.

Bug: 439624904
Test: atest WMShellUnitTests:com.android.wm.shell.splitscreen
Flag: EXEMPT bugfix
Change-Id: I8fb5990fb7eaa6df3068e6ef94e2ee598e0b9a9f