am: Decouple OomAdjuster from ProcessStats
OomAdjuster had a direct dependency on ActivityManagerService to update
the process tracker state in ProcessStats. This coupling was explicitly
marked with a TODO to be resolved.

This CL decouples OomAdjuster from AMS by moving the responsibility for
updating the process tracker state to the OomAdjuster.Callback
implementation within ActivityManagerService. A new 'doingAll' flag is
passed to the onProcStateUpdated callback to preserve the original
conditional logic.

Note that although the order of calling AMS.setProcessTrackerStateLOSP()
is changed, all the state used by the method is not modified between the
calling point. So this CL should be the pure refactor in concept.

Bug: 441408003
Test: m services.core
Test: atest MockingOomAdjusterTests OomAdjusterTests
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I4e80002181725bfceed007e85418e5fade7126ce