Allow touches to go to lock icon in AOD
When the device is transitioning to AOD, before
the display goes to DOZE and is ON ("aod deferment state"),
allow touches to get routed to the device entry icon
if its longpress is enabled to allow for the lock-icon longpress
during this ~4s period to work.

Fixes: 433400058
Flag: com.android.systemui.allow_doze_touches_for_lock_icon
Test: atest NotificationShadeWindowViewControllerTest
Test: atest DozeTouchInteractorTest
Test: manual - enroll UDFPS; reboot device; enter AOD;
immediately try to longpress lock icon as device enters
AOD; observe, long press works. If udfps is not enrolled;
long pressing does nothing.

Change-Id: Ib61c0d46050ed4b1fad47d766ab61dc47e8eeae0