Avoid crashing on uninitialised caches
Some RenderThread instances will attempt to load and store values in the
cache before initialising the cache from `setupGraphicsSupport` /
`setupDiskCache`. In particular we have observed this before the phone
is unlocked, leaving us to believe the behaviour is because the user
partition is not yet decrypted and thus these caches cannot be accessed
anyway. Under the assumption that this is WAI, we need to remove the
fatal logs in these cases to avoid bringing down the device.

We will also add logs to dumpsys which indicate whether the cache is in
use or not to help diagnose any future cases where the cache is exiting
early in this branch unintentionally.

Bug: 269117286
Flag: com.android.graphics.hwui.flags.separate_pipeline_cache
Test: m
Test: boot to home
Test: adb logcat | rg PersistentGraphicsCache
Change-Id: I6c83792ee14d692795594cb9391ec96f0e34d407