BouncyBall: Allow install on release/older devices
This app is focused towards testing and qualifying devices currently
under development built against the latest API.

However, this is also a useful tool for released devices and (mildly)
older releases.

This app does take advantage of a few APIs introduced in API level 36.
We put in workarounds for those so this app can be used on older APIs.
But we draw the line at API level 30 (Android 11 release), when the
Config.getDisplay() API was introduced.

Bug: 441133956
Test: Successfully installed and ran (automated and manually) this app on a release device on API level 35; still able to run and test on devices under development
Flag: EXEMPT TEST_ONLY
Change-Id: I0aa147ea9183f1d3e9a7204f88b076719d313c71