Add preloading attributes for Native App Zygote
This commit adds two new fields, zygotePreloadNativeLib and
zygotePreloadNativeFunc, for specifying the name of the library to be
loaded and name of the function to be called when the library is loaded.

This is done as a part of implementing Native App Zygote
(go/native-app-zygote).

Bug: 442731726
Bug: 442732151
Bug: 405158601
API-Coverage-Bug: 456642738
Test: m
Flag: android.os.native_framework_prototype
Flag: android.os.native_app_zygote
Change-Id: I497896074aa0e56b1fd9c03ca38c284572dc0ad1