Add ls custom clock to dual shade
This reorganizes the nessecary components to render the lockscreen clock
to be easily buildable from other locations, and uses them to render the
relevant custom clock within the notification shade for dual-shade.

Because of the way STL currently handles motion from a nested scene,
opening the shade over the lockscreen transitions but then it
disappears. The element is correctly positioned when opening the shade
from other scenes. This issue is being tracked at b/445740687.

The current version is minimally acceptable, but additional bugs have
been created to complete the nessecary follow-up work. Building the
POR shade transition is tracked by b/445743080. Making the small clock
react to the height of it's container is tracked by b/445741481.

Bug: 440557251
Flag: com.android.systemui.scene_container
Test: Manually verified the clock is visible
Change-Id: I27742605653655b182d893f1120c59df8cdd1f93