[SB][ComposeIcons] Ellipsizing support
This change introduces an EllipsizingIconsRow that
can animate the last visible icon into an overflow dot
when there isn't enough space.

The overflow calculation and animation is achieved in
a single composition thanks to the animateOverflow
modifier. This modifier reads the results of the Layout
phase in the draw phase and animates the transition to
an overflow dot in a custom draw method.

Placement animations will be handled in a followup CL.

Overflow animation recording in b/438171539 comment#2
Test: manual
Test: Tested all transitions in a custom test app.
Bug: 438171539
Flag: com.android.systemui.status_bar_system_status_icons_in_compose
Change-Id: I3231185bf635028af42367778fcd5d48d7a11bb7

Change-Id: I73a30311e1f37635eacaf04a69aeddbe916ffa48