Expose isDeviceManaged state in DevicePolicyManager
In particular, this will be used by the UserManager to allow deleting
the last full admin on managed devices. We expose this method via
DeviceStateCache to ensure that there are no deadlocks when
DevicePolicyManagerService calls UserManagerService, e.g. to remove a
user, which in turn calls DevicePolicyManagerService to check if the
device is managed.

Flag: android.app.admin.flags.multi_user_management_device_provisioning
Test: m services.devicepolicy
Bug: 390162247
Bug: 394970894
API-Coverage-Bug: 442780973

Change-Id: I3a737bb7930f7711ff785ecf6ec9de34dd5b249b