[flexiglass] Better test harness disablement of dual shade educational tooltips
Prior work disabled the showing of dual shade educational tooltips in
test harnessed environemnts (so screenshot tests don't see tooltips) but
the solution was only half-baked: it only worked on the tooltips
view-model level which meant that the shade header view-model was still
seeing that there's a tooltip, causing it to bounce-animate the status
bar chip even though the tooltip wasn't showing.

This CL adds test harness disablement to the second view-model as well.
Now, in test harness mode not only do the tooltips not show but also
their status bar chips don't bounce.

Bug: 445751156
Test: manually verified that the tooltips show up now when not in test
harness mode (which was likely the case before it's just that they
didn't show in test harness mode which caused me to file the bug) and do show
normally when not in test harness mode
Flag: com.android.systemui.scene_container

Change-Id: Ibc8a0d0ec7f4546cec4712cbc1e57243dda5d86c