Prevent scheduling redundant frames for insets animation
Before this CL, when there is any insets animation callback, the host
would apply the insets surface change in the next frame to make sure
the insets surface change is applied with the view changes. However, it
could be redundant to schedule a frame if there is no change of the
view.

This CL only apply the insets surface change with views while there is
any view change or there might be a possible view change (e.g., when it
is ANIMATION_TYPE_USER).

Bug: 438603965
Bug: 446575234
Flag: com.android.window.flags.prevent_scheduling_redundant_frame
Test: atest WindowInsetsAnimationTests
Change-Id: If5886cb2f7376d083a66cb319304c9c37c79daed