Hide brightness slider if the shade display is not internal
The previous logic only used the display id, but as we can now map the
default display to an external physical display (e.g. closed lid case),
it's better to double check the display type instead.

DisplayTypeRepository has been created to extract the logic to monitor a
variable display type onDisplayChanged. In the future, if we have a way
to filter for certain onDisplayChanged, we should do it as the number of
onDisplayChanged is usually high and excessive.

Bug: 444223499
Test: QuickSettingsContainerViewModelTest, ShadeDisplayTypeRepositoryTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I5b1b03dff4ade097de1992dcb1479936fbd1764c