Fix surface animation frame callback deprecation
The original change set the chroreographer provider potentially as null
when going down the deprecation path. The problem with this was that
created a new callback provider using the choreographer on the current
thread which may not have had a Looper associated with it. To address
this, make sure the provider is done on the correct thread.

Flag: com.android.window.flags.deprecate_surface_animation_frame_callback
Bug: 443142083
Test: atest WmTests:SurfaceAnimationRunnerTest
Change-Id: I3ea85a3737328c77ffc0c5b4c94e289d9a6ddf54