[flexiglass] Drive KTF to the currentScene, not the toScene.
These are typically the same scene, but during an interrupted transition from A -> B, where we're going back to A, the scene framework attempts to be smart and drive the existing A -> B transition backwards to A, from progress>0f to progress=0f.

This works for nice UI elements that are normal and can render arbitrary progress values between A/B without caring about direction. However, this does not work for KTF, which needs to immediately begin transitioning.

Bug: 438799387
Fixes: 442873363
Test: atest LockscreenSceneTransitionInteractorTest
Test: launch camera while unlocked and observe KTF no longer thrashes all over the place
Flag: com.android.systemui.scene_container
Change-Id: Ib11be0fb80ce56e3d5119dfaa6efad7a3969e08c