Decouple CE and DE database transactions in AccountManagerService
This change introduces a flag `detachDeCe` to control whether the CE (Credential Encrypted) and DE (Device Encrypted) databases are treated as separate entities. When the flag is enabled, operations on the CE database no longer run within a transaction on the DE database. This allows for more granular control and potentially better performance by avoiding cross-database transactions. The `AccountsDb` class is updated to manage a separate `CeDatabaseHelper` instance and provide methods to interact with the CE database independently.

Bug: 444846497
Flag: com.android.server.accounts.detach_de_ce
Test: CtsAccountManagerTestCases
Change-Id: Id9cdd59fdba5a4183282e0584de132022d692ab6