JobScheduler: Throw exception with debug info when job limit is exceeded
Previously, when an app scheduled more jobs than the allowed limit,
JobSchedulerService would throw an IllegalStateException. For regular apps,
this was sufficient, but when the system UID was the culprit, the generic
exception message made it difficult to identify which system component was
responsible for the excessive job scheduling.

This change enhances the exception message for the system UID to include a list
of the most frequently scheduled jobs. This provides immediate, actionable
information to help developers identify the source of the issue.


Bug: 438864409
Test: atest FrameworksMockingServicesTests:com.android.server.job
Test: atest CtsJobSchedulerTestCases
Flag: com.android.server.job.enhance_system_job_limit_exception

Change-Id: Iec66ac25571bae40b517b3a8e14553da60157196