[DocumentsContract] Add content sync state flags column
Add an optional content sync state flags column to
DocumentsContract.Document. This will expose the content sync
state for each item in DocumentsProvider file system. This
is represented as a bitmask. The valid flags are:
- FLAG_AVAILABLE_LOCALLY
- FLAG_LOCAL_CHANGES
- FLAG_REMOTE_CHANGES
- FLAG_UPLOAD_PROGRESS
- FLAG_DOWNLOAD_PROGRESS
- FLAG_UPLOAD_ERROR
- FLAG_DOWNLOAD_ERROR

If the sync state is unknown or not provided, null will be returned.

Put these changes behind the enable_sync_state flag.

Flag: android.provider.enable_sync_state
Bug: 415947063
Bug: 415947192
Bug: 445029559
Test: atest android.provider.cts.DocumentsContractTest
Change-Id: Ic0efc9ae5926236b564f9ee50392a277c1a5ad6f