MirrorView: Use the compounded alpha of the view ancestors
Use a ViewTreeObserver to listen to any changes in the MirrorView's view
hierarchy, and calculate the compounded alpha of all of the ancestors of
the MirrorView. This is the aplha value that will be applied to the
SurfaceView.

This will make MirrorView behave more similarly to other views in the
hierarchy.

Bug: 441475896
Test: Presubmit, manual with hellocomputercontrol test app
Flag: android.companion.virtualdevice.flags.computer_control_access
Change-Id: Ide8148841ca37bfa2569b5cdfc7fcadb315d5d6b