Clean up window decors for non-transition task vanishes
This is done with both FreeformTaskListener and FullscreenTaskListener.
It doesn't make sense not to do it for RootTaskDesksOrganizers.

Particularly this is useful to clean up window decorations when an
invisible task vanishes, which doesn't trigger transitions. This is also
used in CTS test scenarios where TestTaskOrganizer takes over the task
organizer status from ShellTaskOrganizer, and existing tasks appear to
have vanished. We should clean up window decorations in those cases too.

Bug: 437369884
Test: atest WindowInsetsPolicyTest#testForcedConsumedTopInsets
Test: atest RootTaskDesksOrganizersTest
Flag: EXEMPT BUGFIX
Change-Id: Ie95922a001a9c855255fe12be3f05dab25271edb